pyproject.toml
[tool.poetry]
name = "geoapps"

version = "0.12.0-alpha.1"
license = "MIT"
description = "Open-sourced Applications in Geoscience"
authors = ["Mira Geoscience <dominiquef@mirageoscience.com>"]
repository = "https://github.com/MiraGeoscience/geoapps"
documentation = "https://geoapps.readthedocs.io/en/latest/"
homepage = "https://mirageoscience.com"
readme = "README.rst"

packages = [
    { include = "geoapps" },
    { include = "geoapps-assets" },
]
exclude = [
    "geoapps-assets/**/*.geoh5",
    "geoapps-assets/uijson/**",
]

keywords = ["geology", "geophysics", "earth sciences"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Physics",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Natural Language :: English",
]

[tool.poetry.scripts]
start_notebook = { callable = "geoapps.scripts.start_notebook:main", extras = ["apps"] }

[tool.poetry.dependencies]
python = "^3.10, < 3.11"

## dependencies for a minimal environment
numpy = "~1.23.5"  # also geoh5py, simpeg and geoana
scipy = "~1.10.1"
semver = "~3.0.0"
c-blosc2 = "~2.13.2"

geoh5py = {version = "~0.9.0-alpha.4", source = "pypi", allow-prereleases = true}
#geoh5py = {url = "https://github.com/MiraGeoscience/geoh5py/archive/refs/heads/release/0.9.0.zip#sha256="}
mira-omf = {version = "~3.2.0-alpha.1", source = "pypi", allow-prereleases = true}
#mira-omf = {url = "https://github.com/MiraGeoscience/omf/archive/refs/heads/release/3.2.0.zip#sha256="}
las-geoh5 = { version = "~0.2.0-alpha.1", source = "pypi", allow-prereleases = true}
#las-geoh5 = {url = "https://github.com/MiraGeoscience/las-geoh5/archive/refs/heads/release/0.2.0.zip#sha256="}
param-sweeps = {version = "~0.1.7-alpha.4", source = "pypi", allow-prereleases = true}
#param-sweeps = {url = "https://github.com/MiraGeoscience/param-sweeps/archive/refs/heads/release/0.1.7.zip#sha256="}

## dependencies for core algorithms
dask = {version = "2022.10.*", extras = ["distributed"], optional = true}
discretize = {version = "~0.10.0", optional = true }  # also in simpeg
distributed = {version = "2022.10.*", optional = true}  # because conda-lock doesn't take dask extras into account
fiona = {version = "~1.9.2", optional = true}
fsspec = {version = "2022.*", optional = true }  # also in simpeg
gdal = {version = "~3.6.3", optional = true}
matplotlib = {version = "~3.7.1", optional = true }  # also in simpeg and geoana
pandas = {version = "~1.5.3", optional = true }  # also in simpeg
scikit-learn = {version = "~1.2.2", optional = true }  # also in simpeg
tqdm = {version = "^4.64.0", optional = true }  # also in las-geoh5

#simpeg-archive = {url = "https://github.com/MiraGeoscience/simpeg/archive/refs/heads/feature/simpeg_archive.zip#sha256="}
simpeg-archive = {version = "~0.11.0.dev6", source = "pypi", optional = true}
mira-simpeg = {url = "https://github.com/MiraGeoscience/simpeg/archive/refs/heads/release/v0.19.0.dev8+geoapps.0.12.0.zip#sha256=", optional = true}
#mira-simpeg = {version = "~0.19.0.dev8", source = "pypi", optional = true}

## indirect dependencies, forcing them here for installation through Conda not pip
empymod = {version = "~2.2.1", optional = true}  # from simpeg and geoana
geoana = {version = "~0.4.0", optional = true}  # from simpeg
h5py = "^3.2.1"  # from geoh5py
lasio = "~0.31"  # from las-geoh5
mkl = {version = "2022.*", optional = true }  # from simpeg
Pillow = "^10.0.1"  # from geoh5py
properties = "~0.6.0"  # from omf
pydantic = "~2.5.3"  # from las-geoh5
pydiso = {version = "~0.0.3", optional = true}  # from simpeg
pymatsolver = {version = "~0.2.0", optional = true}  # from simpeg
pypng = "^0.20220715"  # from omf
scikit-image = {version = "~0.19.2", optional = true}
six = "^1.16"  # from omf
vectormath = "~0.2.0"  # from omf
zarr = {version = "~2.14.2", optional = true}  # from simpeg using Dask

## dependencies for UI
ipyfilechooser = {version = "~0.6.0", optional = true}
ipywidgets = {version = "^7.6.5", optional = true}
plotly = {version = "~5.13.1", optional = true}
pyqtwebengine = {version = "~5.15", optional = true}
pyside2 = {version = "~5.15", optional = true}
dash-daq = {version = "~0.5.0", optional = true}

## about pip dependencies
# to be specified to work with conda-lock
# - from PyPI: my_package = { version = "1.2.3", source = "pypi" }
# - from URL:
#   - for a tags:   my_package = { url = "https://github.com/ORGANISATION/REPO/archive/refs/tags/VERSION_TAG.zip#sha256=" }
#   - for a branch: my_package = { url = "https://github.com/ORGANISATION/REPO/archive/refs/heads/BRANCH.zip#sha256=" }
# Note - conda-lock does not support the syntax: my_package = { git = ... }
#        while poetry2conda supports only { git = "...", tag = "..." }, and not { url = ... }

[tool.poetry.group.dev.dependencies]
Pygments = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
tomli = "*"

[tool.poetry.extras]
core = [
    "dask",
    "discretize",
    "distributed",
    "empymod",
    "fiona",
    "fsspec",
    "gdal",
    "geoana",
    "matplotlib",
    "mira-simpeg",
    "mkl",
    "pandas",
    "pydiso",
    "pymatsolver",
    "scikit-image",
    "scikit-learn",
    "simpeg-archive",
    "tqdm",
    "zarr",
]

# extras for apps user interface
apps = [
    # should include "core", but need non-overlapping extra lists for conda-lock
    "dash-daq",
    "ipyfilechooser",
    "ipywidgets",
    "plotly",
    "pyqtwebengine",
    "pyside2",
]

[tool.conda-lock]
platforms = ['win-64', 'osx-64', 'linux-64']
channels = [
    'conda-forge',
    'kinverarity'  # for lasio
]

[tool.conda-lock.dependencies]
libblas = "*=*mkl"

[tool.isort]
# settings for compatibility between ``isort`` and ``black`` formatting
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.black]
# defaults are just fine

[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
scripts_are_modules = true
show_error_context = true
show_column_numbers = true
check_untyped_defs = true

plugins = [
    #    'numpy.typing.mypy_plugin'
]

[tool.pytest.ini_options]
#addopts =

[tool.coverage.run]
branch = true
source = ["geoapps"]
omit = []

[tool.coverage.report]
exclude_lines = [
    "raise NotImplementedError",
    "pass",
    "if TYPE_CHECKING",
    "pragma: no cover"
]


[tool.coverage.html]
skip_empty = true
skip_covered = true

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
